requirements.txt
# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Web Interface
streamlit>=1.28.0
streamlit-image-coordinates>=0.1.5

# Medical Image Processing
pydicom>=2.4.0  # For DICOM file support
opencv-python>=4.8.0

# Data Processing
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Explainability
grad-cam>=1.4.0
captum>=0.6.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.10.0
flake8>=6.1.0
mypy>=1.5.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.26.0
notebook>=7.0.0

# Logging
loguru>=0.7.2

# Image processing
albumentations>=1.3.0  # Advanced image augmentations

